get commitTS from backup meta and separate subcommand for br
